V53-V6 services for GetResourcesByClaimSetIdQuery
